# Source/xrserver/static/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "xr connect and aviation documentaion",
		"version": "1.0.0",
		"title": "XR Connect API"
	},
...
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
		"schemas": {
...
			 "addMediaAzure": {
				"type": "object",
				"properties": {

					"FileURL": {
						"type": "string"
					},
                    "FileName": {
						"type": "string"

					},
                    "MediaType":{
                        "type": "string"

                    },
                    "CompanyName":{
                        "type": "string"

                    },
                    "UploadedBy": {
						"type": "string"
					},
                    "AccessType": {
						"type": "string",
						"enum":["Public","Private"]


					}
				}
			},
...
				"getPrivateMediaList": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					}

				}
			},
...
			"getMediaList": {
				"type": "object",
				"properties": {
					"MediaType": {
						"type": "string"
					}

				}
			},
...
			"deleteContentFile":{
				 "type": "object",
				 "properties": {
					"mediaID": {
						"type": "string"
					}
                }
			},
...
			"getMediaPaginationList":{
				 "type": "object",
				 "properties": {
					"MediaType": {
						"type": "string"
					}
                }
			},
...
		}
	},
	"paths": {
...
		"/media/getMediaList": {
			"post": {
				"tags": [
					"Media"
				],
				"summary": "getMediaList ",
				"requestBody": {
					"description": "getMediaList API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/getMediaList"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getMediaList"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getMediaList"
						}
					}
				}
			}
		},
		"/media/getMediaPaginationList": {
			"post": {
				"tags": [
					"Media"
				],
				"summary": "getMediaPaginationList ",
				"requestBody": {
					"description": "getMediaPaginationList API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/getMediaPaginationList"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getMediaPaginationList"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getMediaPaginationList"
						}
					}
				}
			}
		},
...
        "/media/getPrivateMediaList": {
			"post": {
				"tags": [
					"Media"
				],
				"summary": "getPrivateMediaList ",
				"requestBody": {
					"description": "getPrivateMediaList API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/getPrivateMediaList"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getPrivateMediaList"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getPrivateMediaList"
						}
					}
				}
			}
        },
...
		"/media/deleteContentFile":{
			"post" : {
				"tags": [
					"Media"
				],
				"summary": "deleteContentFile",
				"requestBody": {
					"description": " deleteContentFile API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/deleteContentFile"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/deleteContentFile"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/deleteContentFile"
						}
					}
				}
			}
		},
...
		"/media/addMediaAzure":{
			"post" : {
				"tags": [
					"Media"
				],
				"summary": "addMediaAzure",
				"requestBody": {
					"description": " addMediaAzure API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/addMediaAzure"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/addMediaAzure"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/addMediaAzure"
						}
					}
				}
			}
		},
...
	}
}